ML/source/requirements.txt
torch
transformers
pydub
librosa
fpdf
python-docx # Uncomment this if you need 'docx' 
#flask
#Werkzeug
numpy
#resemblyzer
#pandas
requests
fastapi
pydantic
uvicorn
dotenv #pip install python-dotenv
#comment